requirements.txt
dynaconf==3.1.7
pydantic==1.9.0
bcrypt==3.2.0
pycryptodome==3.14.1
pandas
tabulate
pyinstaller
